requirements.txt
BeautifulSoup==3.2.1
Django==1.5.5
Fabric==1.8.3
South==0.8.4
amqp==1.4.4
anyjson==0.3.3
billiard==3.3.0.16
boto==2.27.0
celery==3.1.10
dj-database-url==0.1.2
django-analytical==0.16.0
django-annoying==0.7.9
django-appconf==0.6
django-celery==3.1.10
django-celery-with-redis==3.0
django-compressor==1.3
django-debug-toolbar==1.0.1
django-extensions==1.3.3
django-heroku-memcacheify==0.5
django-heroku-postgresify==0.3
django-nose==1.2
django-pylibmc-sasl==0.2.4
django-selector==0.4
django-ses==0.6.0
django-staticfiles==1.2.1
django-storages==1.1.8
ecdsa==0.11
gunicorn==18.0
kombu==3.0.14
nose==1.3.1
paramiko==1.12.3
psycopg2==2.5.2
pycrypto==2.6.1
pylibmc==1.2.3
pytz==2014.2
redis==2.9.1
selenium==2.41.0
six==1.6.1
splinter==0.6.0
sqlparse==0.1.11
versiontools==1.9.1
wsgiref==0.1.2
